cake_site/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib
bcrypt==4.3.0
python-multipart
pydantic-settings
python-dotenv
websockets
